# perf: reduce allocs when creating $_SERVER 