--- a/frankenphp.c
+++ b/frankenphp.c
@@ -673,6 +673,8 @@ void frankenphp_register_bulk_variables(go_string known_variables[27],
                                  track_vars_array);
     }
   }
+
+  free(dynamic_variables);
 }
 
 static void frankenphp_register_variables(zval *track_vars_array) {