--- a/frankenphp.c
+++ b/frankenphp.c
@@ -581,10 +581,8 @@ static void frankenphp_register_known_variable(const char *key, go_string value,
   }
 
   size_t new_val_len;
-  // FIXME: data must not be modified, make a copy if the default filter isn't
-  // used
-  if (sapi_module.input_filter(PARSE_SERVER, key, (char **)&value.data,
-                               value.len, &new_val_len)) {
+  if (sapi_module.input_filter(PARSE_SERVER, key, &value.data, value.len,
+                               &new_val_len)) {
     php_register_variable_safe(key, value.data, new_val_len, track_vars_array);
   }
 }
@@ -597,10 +595,8 @@ frankenphp_register_variable_from_request_info(const char *key, char *value,
   }
 
   size_t new_val_len;
-  // FIXME: data must not be modified, make a copy if the default filter isn't
-  // used
-  if (sapi_module.input_filter(PARSE_SERVER, key, (char **)&value,
-                               strlen(value), &new_val_len)) {
+  if (sapi_module.input_filter(PARSE_SERVER, key, &value, strlen(value),
+                               &new_val_len)) {
     php_register_variable_safe(key, value, new_val_len, track_vars_array);
   }
 }
@@ -670,10 +666,8 @@ void frankenphp_register_bulk_variables(go_string known_variables[27],
 
   size_t new_val_len;
   for (size_t i = 0; i < size; i++) {
-    // FIXME: data must not be modified, make a copy if the default filter isn't
-    // used
     if (sapi_module.input_filter(PARSE_SERVER, dynamic_variables[i].var,
-                                 (char **)&dynamic_variables[i].data,
+                                 &dynamic_variables[i].data,
                                  dynamic_variables[i].data_len, &new_val_len)) {
       php_register_variable_safe(dynamic_variables[i].var,
                                  dynamic_variables[i].data, new_val_len,