--- a/frankenphp.c
+++ b/frankenphp.c
@@ -677,9 +677,8 @@ void frankenphp_register_bulk_variables(go_string known_variables[27],
     }
   }
 
-  // Just a test
-  // free(dynamic_variables);
-  // dynamic_variables = NULL;
+  free(dynamic_variables);
+  dynamic_variables = NULL;
 }
 
 static void frankenphp_register_variables(zval *track_vars_array) {