The main idea is to allocate only one time (Go-side) the memory needed to populate `$_SERVER` environment variables.

All the `strings.Clone()` calls are a workaround for https://github.com/golang/go/issues/65286#issuecomment-1920087884, they will not be necessary anymore in Go 1.22.

On my machine, this implementation is around 5% faster than the current one (`BenchmarkServerSuperGlobal`, introduced in #539).

Please note that when running Go benchmarks, the alloc counter doesn't take into account C allocations.