@dunglas 2024-02-01
I did some more benchmarks using the k6 script we provide, and they are non-conclusive (the gap is small and the winner varies on each run...). @withinboredom, would you mind checking if you notice any improvement or deterioration using your benchmarks?

In the meantime, let's mark this patch as a draft.