@dunglas 2024-02-12
With the latest changes, the gains are more significant: 32% fewer allocations and slightly less memory used in Go (probably much less in C but it's hard to measure).

Before:

```
goos: darwin
goarch: arm64
pkg: github.com/dunglas/frankenphp
BenchmarkServerSuperGlobal
BenchmarkServerSuperGlobal-10    	    6571	    178113 ns/op	   19977 B/op	      93 allocs/op
PASS
ok  	github.com/dunglas/frankenphp	2.435s
```

After:

```
goos: darwin
goarch: arm64
pkg: github.com/dunglas/frankenphp
BenchmarkServerSuperGlobal
BenchmarkServerSuperGlobal-10    	    6859	    173809 ns/op	   19955 B/op	      63 allocs/op
PASS
ok  	github.com/dunglas/frankenphp	2.325s
```