@dunglas 2024-02-15
K6 benchmark on a Macbook Pro (M1 Pro):

Before:

```
     execution: local
        script: load-test.js
        output: -

     scenarios: (100.00%) 1 scenario, 100 max VUs, 1m0s max duration (incl. graceful stop):
              * default: 100 looping VUs for 30s (gracefulStop: 30s)


     ✓ is status 200
     ✓ is echoed

     checks.........................: 100.00% ✓ 435614      ✗ 0     
     data_received..................: 910 MB  30 MB/s
     data_sent......................: 1.1 GB  36 MB/s
     http_req_blocked...............: avg=2.1µs    min=0s     med=0s      max=4.59ms  p(90)=1µs     p(95)=1µs    
     http_req_connecting............: avg=1.2µs    min=0s     med=0s      max=3.06ms  p(90)=0s      p(95)=0s     
     http_req_duration..............: avg=13.72ms  min=1.25ms med=13.43ms max=67.75ms p(90)=15.46ms p(95)=16.43ms
       { expected_response:true }...: avg=13.72ms  min=1.25ms med=13.43ms max=67.75ms p(90)=15.46ms p(95)=16.43ms
     http_req_failed................: 0.00%   ✓ 0           ✗ 217807
     http_req_receiving.............: avg=881.68µs min=8µs    med=527µs   max=40.65ms p(90)=1.68ms  p(95)=2.38ms 
     http_req_sending...............: avg=14.61µs  min=8µs    med=12µs    max=3.28ms  p(90)=19µs    p(95)=23µs   
     http_req_tls_handshaking.......: avg=0s       min=0s     med=0s      max=0s      p(90)=0s      p(95)=0s     
     http_req_waiting...............: avg=12.83ms  min=444µs  med=12.67ms max=63.73ms p(90)=14.21ms p(95)=14.82ms
     http_reqs......................: 217807  7256.993086/s
     iteration_duration.............: avg=13.77ms  min=1.36ms med=13.48ms max=67.8ms  p(90)=15.51ms p(95)=16.48ms
     iterations.....................: 217807  7256.993086/s
     vus............................: 100     min=100       max=100 
     vus_max........................: 100     min=100       max=100 


running (0m30.0s), 000/100 VUs, 217807 complete and 0 interrupted iterations
default ✓ [======================================] 100 VUs  30s
```

After:

```
     execution: local
        script: load-test.js
        output: -

     scenarios: (100.00%) 1 scenario, 100 max VUs, 1m0s max duration (incl. graceful stop):
              * default: 100 looping VUs for 30s (gracefulStop: 30s)


     ✓ is status 200
     ✓ is echoed

     checks.........................: 100.00% ✓ 437732      ✗ 0     
     data_received..................: 915 MB  31 MB/s
     data_sent......................: 1.1 GB  36 MB/s
     http_req_blocked...............: avg=2.2µs    min=0s     med=0s      max=4.43ms  p(90)=1µs     p(95)=1µs    
     http_req_connecting............: avg=1.21µs   min=0s     med=0s      max=2.97ms  p(90)=0s      p(95)=0s     
     http_req_duration..............: avg=13.65ms  min=1.33ms med=13.38ms max=60.16ms p(90)=15.36ms p(95)=16.25ms
       { expected_response:true }...: avg=13.65ms  min=1.33ms med=13.38ms max=60.16ms p(90)=15.36ms p(95)=16.25ms
     http_req_failed................: 0.00%   ✓ 0           ✗ 218866
     http_req_receiving.............: avg=878.69µs min=8µs    med=522µs   max=23.27ms p(90)=1.63ms  p(95)=2.32ms 
     http_req_sending...............: avg=14.71µs  min=7µs    med=12µs    max=3.78ms  p(90)=19µs    p(95)=23µs   
     http_req_tls_handshaking.......: avg=0s       min=0s     med=0s      max=0s      p(90)=0s      p(95)=0s     
     http_req_waiting...............: avg=12.76ms  min=305µs  med=12.63ms max=57.49ms p(90)=14.17ms p(95)=14.71ms
     http_reqs......................: 218866  7292.430161/s
     iteration_duration.............: avg=13.7ms   min=1.37ms med=13.43ms max=60.21ms p(90)=15.41ms p(95)=16.31ms
     iterations.....................: 218866  7292.430161/s
     vus............................: 100     min=100       max=100 
     vus_max........................: 100     min=100       max=100 


running (0m30.0s), 000/100 VUs, 218866 complete and 0 interrupted iterations
default ✓ [======================================] 100 VUs  30s
```

(0,5% improvement). Memory usage seems improved too.

It would be nice if someone could try the benchmark on Linux.


I will also try to use some sync.Pool to prevent memory allocations.