@ChrisRiddell 2024-02-19
> (0,5% improvement). Memory usage seems improved too.
> 
> It would be nice if someone could try the benchmark on Linux.

Where about is the k6 file located to run the same benchmark but on linux?