@dunglas 2024-03-05
@maypok86 sorry to bother you again (and I hope that it's not for nothing) but it looks like the latest failure is also related to Otter. It may be this know Go bug: https://pkg.go.dev/sync/atomic#pkg-note-BUG