@maypok86 2024-03-05
> @maypok86 sorry to bother you again (and I hope that it's not for nothing) but it looks like the latest failure is also related to Otter. It may be this known Go bug: https://pkg.go.dev/sync/atomic#pkg-note-BUG

@dunglas Damn, yeah, I completely forgot about that when refactoring and there were no tests on 32-bit archs.

Try the `go get -u github.com/maypok86/otter@dev` version, it seems to pass tests on 32-bit architecture.