@maypok86 2024-03-05
The tests seem to have passed, then I'll create a release with the bug fix now.